requirements.txt
aiogram==2.4
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.7.0
certifi==2019.11.28
chardet==3.0.4
idna==2.8
multidict==4.7.3
pytz==2019.3
yarl==1.4.2
regex
dateparser
pytest
python-Levenshtein
requests